Adding issues panel piping.
Add piping from the AdsPageLoadMetricsObserver for when a heavy ads
issue occurs to pipe a message to the DevTools issue panel.  This
requires going through the RenderFrameHost layer to get to the Impl
under content, as the message will be built up there and sent over.

Bug: 1095617
Change-Id: Ib17c38788bb7ddd0d11fcb12abd2ab9b01f7fd3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2175752
Commit-Queue: Eric Robinson <ericrobinson@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780973}