Fix DCHECK crash in browser_controls_offset_manager.cc
When the browser controls params change, we set the bottom shown ratio
to 0 if the new bottom controls height is 0. This causes us to hit a
DCHECK in SetupAnimation() because we don't early out from
UpdateBrowserControlsState() thinking the ratios need animating.

This CL keeps updating the shown ratio even if the controls height is 0.
We already keep the bottom controls shown ratio "alive" even if the
height is 0 during scrolls and show/hide animations.

Bug: 1094529
Change-Id: Ic8fad2fc0c44ad1b879efdd22e93e2fcd5d356c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248931
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Cr-Commit-Position: refs/heads/master@{#780988}