[WebLayer] Add public API for triggering translate
This adds two new methods on Tab: canTranslate() and showTranslateUi().
These can be used to implement a menu item similar to the "Translate..."
item that shows up in Chrome when clicking the three dots menu.

The logic for initiating manual translation was heavily inspired by
ChromeTranslateClient.

Note: I did not add a C++ equivalent API because we don't currently have
any translate UI for WebLayer on desktop.

Another note: TranslateTest is in the private instrumentation test suite
because it uses TestWebLayer. I wasn't able to find a great way of
testing this using purely the public APIs.

Bug: 1091719
Change-Id: I4837616127ca79040ec132ebafb21832cde64fb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250833
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: anthonyvd <anthonyvd@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780996}