[UI] MenuItemView should use layout constants
This CL is one step in the right direction that replaces using the
hard coded constants in MenuConfig by the LayoutConstants.

It's possible to replace the constants in MenuConfig directly
because they are used in multiple places with different semantics.

For default settings, the net effect of this CL is to use 12 pixels
between the icon and the label instead of 8.

Screenshots:
Before: https://screenshot.googleplex.com/mOhKrpXONNH
After: https://screenshot.googleplex.com/GHpw5rccE0Z

Bug: 1044038
Change-Id: Ie95fa4060e1c2803287372584d4fd536eb631c13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254319
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781014}