[LiteVideo] Implements the blocklist and decider for lite videos.
This change adds the blocklist for the LiteVideo optimziation based on
host for both navigations and rebuffer opt-out/low quality signals. The
parameters for how many opt-out events occur for each host are configurable
via finch.

This change also adds the LiteVideoDecider to determine if LiteVideo
can be applied to a navigation and the parameters that should be used
when throttling media requests, specified by a LiteVideoHint.

A future change will implement a web contents observer and keyed service
so that the LiteVideo optimization can query the decider when the feature
is enabled.

Bug: 1095512,1095511
Change-Id: Ibccdbfedad98a98eac3a7d8b2d756a7917c12784
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247383
Auto-Submit: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: rajendrant <rajendrant@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Commit-Queue: Michael Crouse <mcrouse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781015}