Add LiteMode https image compression infobar
This CL adds an one-time infobar that is shown before https image
compression is triggered for the first time, for LiteMode users in
Android. Image compression should be activated only when the infobar has
been shown already. The infobar should be shown only for https and
non-CCT pages.

Subsequent CLs will enhance the infobar decider logic to not show it
for users who enabled LiteMode recently.

The infobar screenshot is here
https://storage.cloud.google.com/chromium-translation-screenshots/3f7399d55b217cb8223baee1ea1e5631eac9fe0c

Bug: 1090843
Change-Id: I5c02c501c791fc53df8984176773574be2cddb9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236048
Commit-Queue: rajendrant <rajendrant@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: Pavel Yatsuk <pavely@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781017}