[Live Caption] Live Caption works on incognito windows.
If Live Caption is enabled on a profile, and a user opens an incognito
window, live caption should also be enabled in incognito.

Additionally, this CL adds tests to check that Live Caption works on
multiple profiles. There is an independent CaptionController on each
profile. If the user changes the enabled state of Live Caption on one
profile, it should not affect other profiles. The same CaptionController
is used on a profile and its incognito windows.

Bug: 1055150, 1095157
Change-Id: Id4cf7f41fd5f3cfda8f79f7b96edc70a8831fbe8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248740
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Reviewed-by: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781018}