Revert "Fix a test failure due to the change of Chrome new tab page url"
This reverts commit 60536744cd2b82a7abfad6b0637b7de0eec613d4.

Reason for revert: According to
https://chromium-review.googlesource.com/c/chromium/src/+/2242194,
the behavior changes of these two policies don't seem correct.

Comments from Moe: That does not sound correct. chrome://newtab
redirects to chrome-search://local-ntp/local-ntp.html or
chrome://new-tab-page (based on the ntp-webui feature flag)
when Google is the DSP; or redirects to a third-party NTP,
if applicable. In off-the-record profiles (guest and incognito)
there is no directs and the WebUI page is associated with
chrome://newtab. The NTP URL to point to should always remain
chrome://newtab.

Original change's description:
> Fix a test failure due to the change of Chrome new tab page url
>
> Validation:
> Tryjobs, the failure was from another test
>
> Change-Id: I2db090b308c05887485edfef894140c6d89c3b2d
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254705
> Reviewed-by: Owen Min <zmin@chromium.org>
> Commit-Queue: Xiang Ji <jxiang@google.com>
> Cr-Commit-Position: refs/heads/master@{#780527}

TBR=zmin@chromium.org,jxiang@google.com

# Not skipping CQ checks because original CL landed > 1 day ago.

Change-Id: I2b4c27aecac5a12a668083cfb0eee5f6249c5aa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2258674
Commit-Queue: Xiang Ji <jxiang@google.com>
Reviewed-by: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781031}