[content] Do not reset RDHUD for window.open() initial navigation.
At the moment we generate an initial fake navigation for a blank window.open(),
which leads to RenderDocumentHostUserData being destroyed and the data
set when WebContents was created.

Add a logic to prevent this from happening — more specifically, do not
reset per-document data when we see a strange navigation
(DidCommitNavigationInternal with a navigation which is not in a
WaitingForCommit state).

R=alexmos@chromium.org,dcheng@chromium.org,sreejakshetty@chromium.org
CC=bfcache-bugs@chromium.org
BUG=1071232

Change-Id: Ib8db0c3006819f2ea41f7fd8174de8ea426d6db0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235598
Commit-Queue: Alexander Timin <altimin@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Sreeja Kamishetty <sreejakshetty@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781035}