[Demo Mode] Keep the locale variant selected during OOBE.
In OOBE, during enrollment into Demo Mode, the user selects a locale
variant. That selection is stored both in the locale state and the
profile preferences. However, the profile is discarded and a new profile is
always used when a Demo Mode session starts. For this reason, the locale
selected by the user is lost and only the fallback browser locale is
kept. This means that "fr-CA" becomes "fr" for Demo Mode.

This change allows Demo Mode to check for the local state locale after
a new session has started, which is set as the application locale of the
profile for the remained of the session.

Bug: b:158695273
Change-Id: I6c7d3f0508ee1a6bd4078309bb9e6110e8968c9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251138
Commit-Queue: Claudio M <claudiomagni@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781088}