arc-a11y: Don't compute name from descendants for ScrollableContainer
CL:2222068 modified ARC++ focusability and accessibility name
computation.
In that change, we introduced concepts of ScrollableContainer and
ToplevelScrollIem. For ToplevelScrollItems, we compute its a11y name
from descendants if necessary. So we shouldn't do it for
ScrollableContainers, but we're not checking it.

This CL fixes the bug.

Bug: b:152930082
Test: unit_tests --gtest_filter="AXTreeSourceArcTest.*"
Test: manual. Open PlayStore, and open "About this app" dialog from app detail page. Each text nodes gets focus from ChromeVox.
Change-Id: Idb84b30c2461e73d6be5f929546c3af712cf2a83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253681
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Sara Kato <sarakato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781095}