Add Chrome Cleaner commit info to FoilVersionSpec
This keeps a record of which commits each version of the Chrome Cleanup
Tool is built with. Currently this information is stored in tags on the
internal repo, but this approach won't work when the tool is built from
the public repo.

Goes with commit http://shortn/_zVs7MpRCuC in the internal repo that
stores the commit hashes in the protobuf.

R=drubery

Bug: 1004848
Change-Id: I6ae4073f10f571b6b70d35085291b68720898287
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253130
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781171}