Propagate full Error object to unprivileged MediaApp upon SaveCopy error.
Previously when we encountered an error during SaveCopy, we would sometimes just pass the error name and miss out on a lot of useful information. This CL removes that handling and just lets the error object propagate through the app.

Other half of this change is http://cl/317591265.

Bug: b/152809763
Change-Id: I65017f6b5359a8f22a7e9a33952b980f547833d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2256115
Reviewed-by: Bugs Nash <bugsnash@chromium.org>
Commit-Queue: Mitchell McDermott <mcdermottm@google.com>
Cr-Commit-Position: refs/heads/master@{#781195}