Pinning shelf item for drag should not scroll to visible
When drag from app list to shelf is started, the dragged item is pinned
to shelf. Currently, items added due to user action (e.g. pinning) are
scrolled to visible when added. This is undesirable behavior for drag
and drop. The item is initially added at the end of the shelf view,
before moving the item to the hovered index. This means that the shelf
would scroll to the end when starting drag from app list, making it
hard to target the beginning of the shelf view when the shelf does not
fit in the available space.

BUG=1097629

Change-Id: I0f04853bf31a80fd7c90d532fc0f4d5e7c761938
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2259475
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781198}