WebLayer: Adjust infobar behavior for never translate lang/site
Chrome's translate infobar disappears when the user selects the
overflow menu options of never translate language/site. However, this
behavior is contingent on a snackbar then appearing informing the user
of their choice and giving them the ability to undo that choice until
the snackbar disappears (a few seconds). WebLayer does not currently
have snackbars in its infobar implementation. Hence, with the translate
UX eng owner (megjablon@) we made the decision to have the selection
of never translate lang/site *not* cause the infobar to disappear in
WebLayer. The user can then either undo their choice or manually close
the infobar as desired. This CL implements that change, including
ensure that the relevant overflow menu item is checked and that
unchecks of these menu items are properly processed.

UX changes discussed in detail here (internal-only link, apologies):

https://docs.google.com/document/d/1o6Q6vs5_MCcDNXx6mouhylhzQbfgItxbwe-KqNAbTlo/edit#heading=h.2hzrc3fbkgdf

file, which will enable the TranslateCompactInfoBar ForTesting methods
to be removed in production.

Binary-Size: Will follow up by moving JNI boundary to be on test support
Bug: 1093846
Change-Id: I79d4955e4a19d3dc3db1edfc34b63b173d4108d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247024
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781207}