Modify ArcAppIcon to fetch two layer icons.
This is the preparing for the adaptive icon feature. For kUncompressed
and kCompressed, no change, just refactor.

Add a new type kAdaptive for ArcAppIcon to fetch both the foreground and
background image, which will be used for adaptive icons. Currently no
one uses the new type yet. There will be some follow up CLs to change
other code to use the new type kAdaptive, hidden with the flag for M85.

BUG=1083331

Change-Id: I0c078d6d9c821aa810c849bda04875a4a3155a60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249201
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Long Cheng <lgcheng@google.com>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781210}