Use printer status icon elements for dropdown destinations
On CrOS, printer icons with a status badge will be displayed next to
each printer in the dropdown.

The printer status responses are saved in a dictionary accessible by the
dropdown. When the dictionary is updated with a new status, the
printers in the dropdown are triggered to recalculate the color of its
status badge.

Screenshot: http://screen/nOgp4BEL9i7

Bug: 1059607
Change-Id: I4422d628ab6627ab74fa1810dda89cbe19a9acf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238509
Commit-Queue: Gavin Williams <gavinwill@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Bailey Berro <baileyberro@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781232}